Update ds1603l.cpp
Removed static unsigned long start_time = millis();

Added static uint32_t start_time = millis();

+++ esphome/components/ds1603l/ds1603l.cpp
@@ -22,7 +22,7 @@ void Ds1603l::update() {
 
 void Ds1603l::loop() {
   static bool initialized = false;
-  static unsigned long start_time = millis();
+  static uint32_t start_time = millis();
 
   // Ignore invalid data during the first 2 seconds after startup
   if (!initialized && (millis() - start_time < 2000)) {
